docker/aiagent/requirements.txt
langchain-openai
langchain-core
psycopg2-binary
flask
langgraph
langsmith